# ARTEMIS-3049 Reduce live page lookup cost