--- a/artemis-server/src/main/java/org/apache/activemq/artemis/core/paging/cursor/PageCursorProvider.java
+++ b/artemis-server/src/main/java/org/apache/activemq/artemis/core/paging/cursor/PageCursorProvider.java
@@ -33,7 +33,7 @@ public interface PageCursorProvider {
 
    PagedReference newReference(PagePosition pos, PagedMessage msg, PageSubscription sub);
 
-   void addPageCache(PageCache cache);
+   void addLivePageCache(LivePageCache cache);
 
    /**
     * @param queueId The cursorID should be the same as the queueId associated for persistence