--- a/artemis-server/src/main/java/org/apache/activemq/artemis/core/paging/cursor/PagePosition.java
+++ b/artemis-server/src/main/java/org/apache/activemq/artemis/core/paging/cursor/PagePosition.java
@@ -34,8 +34,6 @@ public interface PagePosition extends Comparable<PagePosition> {
 
    void setPersistentSize(long persistentSize);
 
-   PagePosition nextMessage();
-
    PagePosition nextPage();
 
 }