https://issues.apache.org/jira/browse/ARTEMIS-3049