@franz1981 2021-01-04
@qihongxu Hi!! 
I've tried to address/improve what you've shown in your comment on https://github.com/apache/activemq-artemis/pull/2494#issuecomment-455086939
Any chances you can test it with that same original test?
I'm going to provide a micro-bench just on the append only list, but only a end-2-end test can shows if the improvements are effective as expected