@franz1981 2021-01-04
Despite my expectations, the results are NOT promising...

This PR
1 thread:
```
Benchmark                                            (chunkSize)  (size)   Mode  Cnt          Score         Error  Units
ConcurrentAppendOnlyChunkListLookupBenchmark.lookup           32    1000  thrpt   16  108968555.072 ±  135438.271  ops/s
ConcurrentAppendOnlyChunkListLookupBenchmark.lookup           32   10000  thrpt   16  105613840.728 ±  475604.647  ops/s
```
3 threads:
```
Benchmark                                            (chunkSize)  (size)   Mode  Cnt         Score          Error  Units
ConcurrentAppendOnlyChunkListLookupBenchmark.lookup           32    1000  thrpt   16  91693844.669 ± 29399345.167  ops/s
ConcurrentAppendOnlyChunkListLookupBenchmark.lookup           32   10000  thrpt   16  78520201.713 ±  3428526.469  ops/s
```

`master`

1 thread:
```
Benchmark                                            (chunkSize)  (size)   Mode  Cnt         Score         Error  Units
ConcurrentAppendOnlyChunkListLookupBenchmark.lookup           32    1000  thrpt   16  59747234.340 ± 1117878.317  ops/s
ConcurrentAppendOnlyChunkListLookupBenchmark.lookup           32   10000  thrpt   16   7699916.690 ±   36448.254  ops/s
```
3 threads:
```
Benchmark                                            (chunkSize)  (size)   Mode  Cnt          Score         Error  Units
ConcurrentAppendOnlyChunkListLookupBenchmark.lookup           32    1000  thrpt   16  181201564.143 ± 2457478.235  ops/s
ConcurrentAppendOnlyChunkListLookupBenchmark.lookup           32   10000  thrpt   16   23011863.725 ±  110276.468  ops/s
```

It shows that 3 consumers won't get the expected speedup and I suppose that's due to the last accessed buffer miss.