@franz1981 2021-01-04
I've added @Contended to some fields of the list + `--jvmArgs="-XX:-RestrictContended"` and now it seems to work better: it seems that the modifications on cached last accessed buffer have a false sharing with lastIndex accesses.