@franz1981 2021-01-04
I've chosen to remove the caching of the last accessed buffer because multiple consumers would scale negatively due to this and instead, let a reloaded live cache to benefit from O(1) accesses instead, while getting a more compacted data representation instead.

The PR is providing a JMH bench to track perf regression on lookup scaling.