@michaelandrepearce 2021-01-09
> I've added `@Contended` to some fields of the list + `--jvmArgs="-XX:-RestrictContended"` and now it seems to work better: it seems that the modifications on cached last accessed buffer have a false sharing with lastIndex accesses.
> 

Im not seeing this flag being altered/added at all in the commits?