@franz1981 2021-01-09
> Im not seeing this flag being altered/added at all in the commits?

Correct: I have used it in a run with JMH while I was trying to improve the page lookup cost of the chunked list, but I have decided that such change was too invasive, see https://github.com/apache/activemq-artemis/pull/3393#issuecomment-754085723